Private doesn't do anything for mixins
It looks like you can't mixin methods as private in ruby. What?! So
anything that is brought in is not able to be hidden from users.

This will cause a problem with the default to use the anything() matcher
in the has_attribute() matcher. For the time being the anything() method
will have to be available from the Ramcrest::HasAttribute module.