Allow such_that matchers to co-exist
Because the such_that matcher was adding the passed block as the
matches? method to its own class, different such_that matchers would end
up stomping on each other's matches? methods.

The matches? method is now put on the singleton class so that each
instance of a such_that matcher is independent of the others.