Allow chaining of match results
This provides two new methods on match results: #or_else and #and_also.

The #or_else method allows for converting mismatches into something
else, which is useful when matching sub-components of larger objects.

The #and_also method allows chaining a sequence of matches together and
only continuing in the case of success. It stops on the first mismatch.