Improve the descriptions for #is
The #is matcher was putting extraneous markers in when it was composed
with other matchers. The markers were supposed to be there for when it
needed to show the boundaries of a value. Since the description of other
matchers should take care of that there is no need for the markers when
it is just showing other matcher's descriptions.